Hibernate is now happy with the schema...less some errors

+++ src/main/java/reviewbot/controllers/BookController.java
@@ -71,7 +71,7 @@ public class BookController {
             return book;
         }
 
-        book = _bookDAO.readOne(new Long(Integer.parseInt(idStr)));
+        book = _bookDAO.readOne(new Integer(Integer.parseInt(idStr)));
         return book;
 
     }
