+++ src/main/java/reviewbot/entity/GenreMap.java
@@ -19,77 +19,77 @@ public class GenreMap implements Serializable{
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
 
-    private Long id;
+    private Integer id;
 
-    private Long book_id;
-    private Long genre_id;
-    private Long subgenre_id;
-    private Long theme_id;
-    private Long misc_id;
-    private Long format_id;
-    private Long award_id;
+    private Integer book_id;
+    private Integer genre_id;
+    private Integer subgenre_id;
+    private Integer theme_id;
+    private Integer misc_id;
+    private Integer format_id;
+    private Integer award_id;
 
-    public Long getAward_id() {
+    public Integer getAward_id() {
         return award_id;
     }
 
-    public void setAward_id(Long award_id) {
+    public void setAward_id(Integer award_id) {
         this.award_id = award_id;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+   public void setId(Integer id) {
         this.id = id;
     }
 
-    public Long getBook_id() {
+    public Integer getBook_id() {
         return book_id;
     }
 
-    public void setBook_id(Long book_id) {
+    public void setBook_id(Integer book_id) {
         this.book_id = book_id;
     }
 
-    public Long getGenre_id() {
+    public Integer getGenre_id() {
         return genre_id;
     }
 
-    public void setGenre_id(Long genre_id) {
+    public void setGenre_id(Integer genre_id) {
         this.genre_id = genre_id;
     }
 
-    public Long getSubgenre_id() {
+    public Integer getSubgenre_id() {
         return subgenre_id;
     }
 
-    public void setSubgenre_id(Long subgenre_id) {
+    public void setSubgenre_id(Integer subgenre_id) {
         this.subgenre_id = subgenre_id;
     }
 
-    public Long getTheme_id() {
+    public Integer getTheme_id() {
         return theme_id;
     }
 
-    public void setTheme_id(Long theme_id) {
+    public void setTheme_id(Integer theme_id) {
         this.theme_id = theme_id;
     }
 
-    public Long getMisc_id() {
+    public Integer getMisc_id() {
         return misc_id;
     }
 
-    public void setMisc_id(Long misc_id) {
+    public void setMisc_id(Integer misc_id) {
         this.misc_id = misc_id;
     }
 
-    public Long getFormat_id() {
+    public Integer getFormat_id() {
         return format_id;
     }
 
-    public void setFormat_id(Long format_id) {
+    public void setFormat_id(Integer format_id) {
         this.format_id = format_id;
     }
